cannot delete categories on derived variables

+++ scrunch/categories.py
@@ -37,6 +37,9 @@ class Category(ReadOnly):
         return dct
 
     def delete(self):
+        if self.resource.body.get('derivation'):
+            raise TypeError("Cannot delete categories on derived variables. Re-derive with the appropriate expression")
+
         return_status = self.resource.delete()
         self.resource.refresh()
         return return_status
